+++ src/main/java/dk/aau/p4/abaaja/Lib/Symbols/TypeDescriptors/MctlTypeDescriptor.java
@@ -1,6 +1,6 @@
 package dk.aau.p4.abaaja.Lib.Symbols.TypeDescriptors;
 
-public abstract class TypeDescriptor {
+public abstract class MctlTypeDescriptor {
     public abstract String get_type_literal();
     public abstract void set_type_literal(String typeLiteral);
 }
